+++ src/mqtt_client.rs
@@ -5,27 +5,35 @@ use crate::{
     Property, {debug, error, info},
 };
 
-use core::{cell::RefCell, str::FromStr};
+use core::{cell::RefCell, convert::TryFrom, str::FromStr};
 
 use embedded_nal::{nb, IpAddr, Mode, SocketAddr};
+use embedded_time::{duration::Seconds, fixed_point::FixedPoint, Clock, Instant};
 
 use generic_array::{ArrayLength, GenericArray};
 use heapless::String;
 
 /// A client for interacting with an MQTT Broker.
-pub struct MqttClient<T, N>
+pub struct MqttClient<T, N, C>
 where
     T: ArrayLength<u8>,
     N: embedded_nal::TcpStack,
+    C: Clock,
+    u32: TryFrom<C::T>,
 {
     /// The network stack originally provided to the client.
     pub network_stack: N,
+    pub clock: C,
 
     socket: RefCell<Option<N::TcpSocket>>,
-    state: RefCell<SessionState>,
+    state: RefCell<SessionState<C>>,
     packet_reader: PacketReader<T>,
     transmit_buffer: RefCell<GenericArray<u8, T>>,
     connect_sent: bool,
+    ping_response_timeout: Seconds<u32>,
+
+    /// Timestamp of the last connection attempt to the broker (for rate-limiting).
+    last_socket_connect_time: Option<Instant<C>>,
 }
 
 /// The quality-of-service for an MQTT message.
@@ -76,10 +84,12 @@ impl<E> From<E> for Error<E> {
     }
 }
 
-impl<T, N> MqttClient<T, N>
+impl<T, N, C> MqttClient<T, N, C>
 where
     T: ArrayLength<u8>,
     N: embedded_nal::TcpStack,
+    C: Clock,
+    u32: TryFrom<C::T>,
 {
     /// Construct a new MQTT client.
     ///
@@ -91,13 +101,12 @@ where
     ///
     /// # Returns
     /// An `MqttClient` that can be used for publishing messages and subscribing to topics.
-    pub fn new(broker: IpAddr, client_id: &str, network_stack: N) -> Result<Self, Error<N::Error>> {
-        // Connect to the broker's TCP port.
-        let socket = network_stack.open(Mode::NonBlocking)?;
-
-        // Next, connect to the broker over MQTT.
-        let socket = network_stack.connect(socket, SocketAddr::new(broker, 1883))?;
-
+    pub fn new(
+        broker: IpAddr,
+        client_id: &str,
+        network_stack: N,
+        clock: C,
+    ) -> Result<Self, Error<N::Error>> {
         let session_state = SessionState::new(
             broker,
             String::from_str(client_id).or(Err(Error::ProvidedClientIdTooLong))?,
@@ -105,15 +114,16 @@ where
 
         let mut client = MqttClient {
             network_stack: network_stack,
-            socket: RefCell::new(Some(socket)),
+            socket: RefCell::new(None),
             state: RefCell::new(session_state),
             transmit_buffer: RefCell::new(GenericArray::default()),
             packet_reader: PacketReader::new(),
             connect_sent: false,
+            clock: clock,
+            ping_response_timeout: Seconds(2),
+            last_socket_connect_time: None,
         };
-
         client.reset()?;
-
         Ok(client)
     }
 
@@ -132,9 +142,10 @@ where
         Ok(read)
     }
 
-    fn write(&self, buf: &[u8]) -> Result<(), Error<N::Error>> {
+    fn write(&self, state: &mut SessionState<C>, buf: &[u8]) -> Result<(), Error<N::Error>> {
         let mut socket_ref = self.socket.borrow_mut();
         let mut socket = socket_ref.take().unwrap();
+        state.last_write_time.replace(self.clock.try_now().unwrap());
         let written = nb::block!(self.network_stack.write(&mut socket, &buf))?;
 
         // Put the socket back into the option.
@@ -172,7 +183,7 @@ where
         let packet = serialize::subscribe_message(&mut buffer, topic, packet_id, properties)
             .map_err(|e| Error::Protocol(e))?;
 
-        match self.write(packet) {
+        match self.write(&mut state, packet) {
             Ok(_) => {
                 info!("Subscribing to `{}`: {}", topic, packet_id);
                 state
@@ -239,12 +250,16 @@ where
         let mut buffer = self.transmit_buffer.borrow_mut();
         let packet = serialize::publish_message(&mut buffer, topic, data, properties)
             .map_err(|e| Error::Protocol(e))?;
-        self.write(packet)
+        self.write(&mut self.state.borrow_mut(), packet)
     }
 
     fn socket_is_connected(&self) -> Result<bool, N::Error> {
         let mut socket_ref = self.socket.borrow_mut();
-        let socket = socket_ref.take().unwrap();
+        let socket = socket_ref.take();
+        if socket.is_none() {
+            return Ok(false);
+        }
+        let socket = socket.unwrap();
 
         let connected = self.network_stack.is_connected(&socket)?;
 
@@ -261,6 +276,15 @@ where
         Ok(())
     }
 
+    /// Non-graceful socket shutdown (will cause broker to send out LWTT if configured,
+    /// etc.).
+    fn shutdown_socket(&mut self) -> Result<(), Error<N::Error>> {
+        self.network_stack
+            .close(self.socket.borrow_mut().take().unwrap())?;
+        self.reset()?;
+        Ok(())
+    }
+
     fn connect_to_broker(&mut self) -> Result<(), Error<N::Error>> {
         self.reset()?;
 
@@ -268,7 +292,7 @@ where
         let packet = serialize::connect_message(
             &mut buffer,
             self.state.borrow().client_id.as_str().as_bytes(),
-            self.state.borrow().keep_alive_interval,
+            *self.state.borrow().keep_alive_interval.integer() as u16,
             &[Property::MaximumPacketSize(
                 self.packet_reader.maximum_packet_length() as u32,
             )],
@@ -276,7 +300,7 @@ where
         .map_err(|e| Error::Protocol(e))?;
 
         info!("Sending CONNECT");
-        self.write(packet)?;
+        self.write(&mut self.state.borrow_mut(), packet)?;
 
         self.connect_sent = true;
 
@@ -291,8 +315,8 @@ where
     where
         for<'a> F: FnMut(&Self, &'a str, &[u8], &[Property<'a>]),
     {
-        let mut state = self.state.borrow_mut();
-        if !state.connected {
+        if !self.state.borrow().connected {
+            let mut state = self.state.borrow_mut();
             if let ReceivedPacket::ConnAck(acknowledge) = packet {
                 if acknowledge.reason_code != 0 {
                     return Err(Error::Failed(acknowledge.reason_code));
@@ -315,7 +339,7 @@ where
                                 String::from_str(id).or(Err(Error::ProvidedClientIdTooLong))?;
                         }
                         Property::ServerKeepAlive(keep_alive) => {
-                            state.keep_alive_interval = keep_alive;
+                            state.keep_alive_interval = Seconds(keep_alive as u32);
                         }
                         _prop => info!("Ignoring property: {:?}", _prop),
                     };
@@ -346,6 +370,7 @@ where
             }
 
             ReceivedPacket::SubAck(subscribe_acknowledge) => {
+                let mut state = self.state.borrow_mut();
                 match state
                     .pending_subscriptions
                     .iter()
@@ -365,27 +390,96 @@ where
                 Ok(())
             }
 
+            ReceivedPacket::PingResp => {
+                self.state.borrow_mut().pending_pingreq_time.take();
+                Ok(())
+            }
+
             _ => Err(Error::Unsupported),
         }
     }
 
-    fn connect_socket(&mut self) -> Result<(), Error<N::Error>> {
-        let mut socket_ref = self.socket.borrow_mut();
-        let socket = socket_ref.take().unwrap();
-
-        // Connect to the broker's TCP port with a new socket.
-        // TODO: Limit the time between connect attempts to prevent network spam.
-        let socket = self
-            .network_stack
-            .connect(socket, SocketAddr::new(self.state.borrow().broker, 1883))?;
-
-        // Store the new socket for future use.
-        socket_ref.replace(socket);
-
+    fn open_socket(&mut self) -> Result<(), N::Error> {
+        let socket = self.network_stack.open(Mode::NonBlocking)?;
+        self.socket.borrow_mut().replace(socket);
         Ok(())
     }
 
-    /// Check the MQTT interface for available messages.
+    fn connect_socket(&mut self) -> Result<(), Error<N::Error>> {
+        // Don't try connecting more than once per second to avoid network spam if
+        // the broker port is closed.
+        if let Some(last_connect) = &self.last_socket_connect_time {
+            let now = self.clock.try_now().unwrap();
+            if now
+                .checked_duration_since(last_connect)
+                .and_then(|diff| Seconds::<u32>::try_from(diff).ok())
+                .map(|diff| diff < Seconds(1u32))
+                .unwrap_or(false)
+            {
+                return Err(Error::Disconnected);
+            }
+        }
+        self.last_socket_connect_time = Some(self.clock.try_now().unwrap());
+
+        // Connect to the broker's TCP port.
+        let addr = SocketAddr::new(self.state.borrow().broker, 1883);
+        info!("Connecting to {}", addr);
+        let mut socket_ref = self.socket.borrow_mut();
+        let socket = socket_ref.take().unwrap();
+        socket_ref.replace(self.network_stack.connect(socket, addr)?);
+        Ok(())
+    }
+
+    fn is_ping_request_due(&self) -> bool {
+        let s = self.state.borrow();
+        if s.keep_alive_interval == Seconds(0_u32) {
+            return false;
+        }
+        // Note(unwrap): We should have at least sent out the initial connect packet.
+        let last_write_time = &s.last_write_time.unwrap();
+        // On error (clock wrapped) be conservative and send out a keepalive ping.
+        let now = self.clock.try_now().unwrap();
+        now.checked_duration_since(last_write_time)
+            .and_then(|diff| Seconds::<u32>::try_from(diff).ok())
+            .map(|diff| diff >= s.keep_alive_interval)
+            .unwrap_or(true)
+    }
+
+    fn send_ping_request(&mut self) -> Result<(), Error<N::Error>> {
+        debug!("Sending keepalive ping request");
+        let mut buffer = self.transmit_buffer.borrow_mut();
+        let packet = serialize::ping_req_message(&mut buffer).map_err(|e| Error::Protocol(e))?;
+        self.write(&mut self.state.borrow_mut(), packet)?;
+        self.state
+            .borrow_mut()
+            .pending_pingreq_time
+            .replace(self.clock.try_now().unwrap());
+        Ok(())
+    }
+
+    fn ping_response_is_overdue(&self) -> bool {
+        let s = self.state.borrow();
+        match &s.pending_pingreq_time {
+            None => false,
+            Some(sent) => {
+                // Be conservative on error (clock wrapped) and just wait until the next ping
+                // to detect a lost connection.
+                let now = self.clock.try_now().unwrap();
+                now.checked_duration_since(sent)
+                    .and_then(|diff| Seconds::<u32>::try_from(diff).ok())
+                    .map(|diff| diff >= self.ping_response_timeout)
+                    .unwrap_or(false)
+            }
+        }
+    }
+
+    /// Check the MQTT interface for available messages, and send a keepalive ping if
+    /// required.
+    ///
+    /// Should be called at least once per second, as per the configured Clock, for
+    /// timely handling of keepalive pings. (More infrequent calls can also cause
+    /// pings to be missed entirely when the keepalive interval is set to a large value
+    /// and the clock wraps around.)
     ///
     /// # Args
     /// * `f` - A closure to process any received messages. The closure should accept the client,
@@ -397,10 +491,20 @@ where
         // If the socket is not connected, we can't do anything.
         if self.socket_is_connected()? == false {
             let result = if self.connect_sent {
+                // Properly close the TCP connection before trying to reconnect. Note
+                // that self.socket can also be present if we've just started a connection
+                // attempt in a recent call to poll (in which case we don't want to close
+                // it).
+                if let Some(socket) = self.socket.borrow_mut().take() {
+                    self.network_stack.close(socket)?;
+                }
                 Err(Error::Disconnected)
             } else {
                 Ok(())
             };
+            if self.socket.borrow().is_none() {
+                self.open_socket()?;
+            }
             self.reset()?;
             self.connect_socket()?;
             return result;
@@ -445,6 +549,19 @@ where
                     .pop_packet()
                     .map_err(|e| Error::Protocol(e))?;
             }
+        }
+
+        if self.ping_response_is_overdue() {
+            error!(
+                "Expected ping response, but none received; dropping broker \
+                    connection (timeout: {})",
+                self.ping_response_timeout
+            );
+            self.shutdown_socket()?;
+        }
+
+        if self.is_ping_request_due() {
+            self.send_ping_request()?;
         }
 
         Ok(())